Extracted functions from mixins module to  a ne helpers module.
